requirements.txt
Flask==0.10.1
Flask-Admin==1.0.8
Flask-Assets==0.10
Flask-Login==0.2.11
Flask-Mail==0.9.0
Flask-Moment==0.4.0
Flask-Principal==0.4.0
Flask-Restless==0.14.0
Flask-SQLAlchemy==1.0
Flask-Security==1.7.3
Flask-WTF==0.10.0
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
SQLAlchemy==0.9.7
Unidecode==0.04.16
WTForms==2.0.1
Werkzeug==0.9.6
alembic==0.6.6
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.18
blinker==1.3
celery==3.1.13
factory-boy==2.4.1
gnureadline==6.3.3
gunicorn==19.1.0
ipython==2.2.0
itsdangerous==0.24
kombu==3.0.21
mimerender==0.5.4
passlib==1.6.2
psycopg2==2.5.3
py-bcrypt==0.4
python-dateutil==2.2
python-mimeparse==0.1.4
pytz==2014.4
raven==5.0.0
redis==2.10.2
requests==2.3.0
simplejson==3.6.2
six==1.7.3
slugify==0.0.1
uWSGI==2.0.6
underscore.py==0.1.6
webassets==0.10.1
wsgiref==0.1.2
